requirements.txt
flask>=2.3.3
gunicorn>=21.0.0
python-dotenv>=1.0.0
serpapi>=0.1.5
web3>=7.10.0
requests>=2.32.3
httpx>=0.27.0
blockrun-llm>=0.2.0
py-clob-client>=0.17.0
google-cloud-storage>=2.10.0